Added reentry and plotting improvements
First stage can now reenter through landing or crash. Second stage can only reenter by burn up to crash.

Due to the small scale in terms of distance for the first stage the user is now asked in the terminal if a plot should be shown for the first stage only.